move some ccure classes away from base
