+++ acslib/ccure/__init__.py
@@ -1,4 +1,5 @@
 from acslib.base import ACSRequestResponse
 from acslib.base.connection import ACSRequestData, ACSRequestMethod
-from acslib.ccure.base import CcureAPI, CcureACS, CCurePersonnel, CCureClearance, CCureCredential
-from acslib.ccure.search import ClearanceFilter, PersonnelFilter
+from acslib.ccure.base import CcureACS
+from acslib.ccure.crud import CcureAPI, CCurePersonnel, CCureClearance, CCureCredential
+from acslib.ccure.search import ClearanceFilter, PersonnelFilter, CredentialFilter
